Correctly report :error values that are not exceptions (#58)
Right now, this:

    Rollbax.report(:error, term, ...)

raises an error when "term" is not an exception. However, this can happen quite
easily: for example, when catching something in a try/catch construct, errors
are not normalized so something like "{:badmap, term}" is not normalized to a
BadMapError exception. With this commit, we stop enforcing exceptions when we
report :error and we normalize errors later on.