switch to a regex match on method frame
The way a stack traces displays a method name seems to have changed between
Elixir version 1.4 and 1.5. Method names that contain spaces will now be
wrapped in quotes, but they were not in 1.4. This causes two tests to fail
because they are expected quotes that don't exist in Elixir versions prior to
1.5. Using this regex, we make the quotes optional to support multiple elixir
versions.